refactor test utilities
