+++ tests/org.eclipse.n4js.ui.tests/src/org/eclipse/n4js/dirtystate/CanLoadFromDescriptionPluginUITest.xtend
@@ -21,6 +21,7 @@ import org.junit.Ignore
 import org.junit.Test
 
 import static org.junit.Assert.*
+import org.eclipse.n4js.tests.util.EclipseUIUtils
 
 /**
  * 
@@ -160,7 +161,7 @@ class CanLoadFromDescriptionPluginUITest extends AbstractCanLoadFromDescriptionT
 	def void test_inEditor_allFromIndex() {
 		prepare("TestInEditorSimple")
 
-		val page = getActivePage()
+		val page = EclipseUIUtils.getActivePage()
 		val editorA1 = openAndGetXtextEditor(fileA1, page)
 		val errorsA1 = getEditorValidationErrors(editorA1)
 		assertEquals("editor for file A should not have any errors", #[], errorsA1)
@@ -178,7 +179,7 @@ class CanLoadFromDescriptionPluginUITest extends AbstractCanLoadFromDescriptionT
 	def void test_inEditor_a1_c() {
 		prepare("TestInEditorA1_C")
 
-		val page = getActivePage()
+		val page = EclipseUIUtils.getActivePage()
 		val editorA1 = openAndGetXtextEditor(fileA1, page)
 		val editorC = openAndGetXtextEditor(fileC, page)
 		val errorsA1 = getEditorValidationErrors(editorA1)
@@ -206,7 +207,7 @@ class CanLoadFromDescriptionPluginUITest extends AbstractCanLoadFromDescriptionT
 	def void test_inEditor_a2_d() {
 		prepare("TestInEditorA2_D")
 
-		val page = getActivePage()
+		val page = EclipseUIUtils.getActivePage()
 		val editorA2 = openAndGetXtextEditor(fileA2, page)
 		val editorD = openAndGetXtextEditor(fileD, page)
 		val errorsA2 = getEditorValidationErrors(editorA2)
@@ -239,7 +240,7 @@ class CanLoadFromDescriptionPluginUITest extends AbstractCanLoadFromDescriptionT
 	def void test_inEditor_a1_b1() {
 		prepare("TestInEditorA1_B1")
 
-		val page = getActivePage()
+		val page = EclipseUIUtils.getActivePage()
 		val editorA1 = openAndGetXtextEditor(fileA1, page)
 		val editorB1 = openAndGetXtextEditor(fileB1, page)
 		val errorsA1 = getEditorValidationErrors(editorA1)
@@ -266,7 +267,7 @@ class CanLoadFromDescriptionPluginUITest extends AbstractCanLoadFromDescriptionT
 	def void test_inEditor_a2_p() {
 		prepare("TestInEditorA2_P")
 
-		val page = getActivePage()
+		val page = EclipseUIUtils.getActivePage()
 		val editorA2 = openAndGetXtextEditor(fileA2, page)
 		val editorP = openAndGetXtextEditor(fileP, page)
 		val errorsA2 = getEditorValidationErrors(editorA2)
@@ -294,7 +295,7 @@ class CanLoadFromDescriptionPluginUITest extends AbstractCanLoadFromDescriptionT
 	def void test_inEditor_a2_q() {
 		prepare("TestInEditorA2_Q")
 
-		val page = getActivePage()
+		val page = EclipseUIUtils.getActivePage()
 		val editorA2 = openAndGetXtextEditor(fileA2, page)
 		val editorQ = openAndGetXtextEditor(fileQ, page)
 		val errorsA2 = getEditorValidationErrors(editorA2)
